% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfBind}
\alias{dfBind}
\title{Row bind data.frames}
\usage{
dfBind(dfList)
}
\arguments{
\item{dfList}{A list of data.frames.}
}
\description{
Helper function for row binding data.frames with diffent columns.
}
\keyword{internal}
