% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dropsMarkers}
\alias{dropsMarkers}
\title{Drops data 2}
\format{An object of class \code{data.frame} with 246 rows and 41723 columns.}
\usage{
dropsMarkers
}
\description{
Drops data 2
}
\keyword{datasets}
