% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dropsPheno}
\alias{dropsPheno}
\title{Drops data 3}
\format{An object of class \code{data.frame} with 7390 rows and 14 columns.}
\usage{
dropsPheno
}
\description{
Drops data 3
}
\keyword{datasets}
