% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA.R
\name{EMMA}
\alias{EMMA}
\title{Compute REML estimates of variance components using EMMA algorithm.}
\usage{
EMMA(dat, trait, K, covar = NULL, nGrids = 100, lLim = -10,
  uLim = 10, eps = .Machine$double.eps^0.25)
}
\arguments{
\item{dat}{A data.frame containing the phenotypic data on which the 
analysis is performed.}

\item{trait}{A trait for which to estimate variance components. This can be
either numeric index or character name of a column in \code{pheno}.}

\item{K}{An optional kinship matrix. If \code{NULL} then matrix
\code{kinship} in \code{gData} is used. If both \code{K} is provided and
\code{gData} contains a matrix \code{kinship} then \code{K} is used.}

\item{covar}{A data.frame of covariates taken into account when
estimating variance components. If \code{NULL} no covariates are used.}

\item{nGrids}{An integer indicating the number of intervals used for local
optimisation within the algorithm.}

\item{lLim}{A numerical value indicating the lower limit of the interval over
which optimisating is done.}

\item{uLim}{A numerical value indicating the upper limit of the interval over
which optimisating is done.}

\item{eps}{A numerical value used as computational tolerance in the
algorithm.}
}
\value{
A list with two components:
\itemize{
\item{\code{varcomp} a vector of genetic variance Vg and residual variance
Ve}
\item{\code{vcovMatrix} The variance covariance matrix corresponding to
the computed variances.}
}
}
\description{
Using the EMMA algorithm as in Kang et al. (2008) to compute REML estimates
of genetic and residual variance components.
}
\references{
Kang et al. (2008) (Efficient Control of Population Structure in
Model Organism Association Mapping. Genetics, March 2008, Vol. 178, no. 3,
p. 1709-1723
}
\keyword{internal}
