% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emmaEigenR}
\alias{emmaEigenR}
\title{EMMA helper functions}
\usage{
emmaEigenR(k, x, eigVals, eigVecs)
}
\arguments{
\item{X}{a q x n covariate matrix, q being the number of covariates and n
being the number of genotypes. q has to be at least one (typically an
intercept).}
}
\description{
Helper functions for computing REML estimates of genetic and residual
variance components using the EMMA algorithm.
}
\keyword{internal}
