% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASHelp.R
\name{estVarComp}
\alias{estVarComp}
\title{Estimate variance components in single trait GWAS}
\usage{
estVarComp(GLSMethod, remlAlgo, trait, pheno, covar, K, chrs, KChr,
  nonMiss, nonMissRepId)
}
\description{
Helper function for estimating variance components in single trait GWAS.
}
\keyword{internal}
