% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASHelp.R
\name{estVarCompNR}
\alias{estVarCompNR}
\title{Helper function for estimating variance components using NR method.}
\usage{
estVarCompNR(dat, fixed, K, nonMiss, nonMissRepId)
}
\description{
Helper function for estimating variance components using NR method.
}
