% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASHelp.R
\name{exclMarkers}
\alias{exclMarkers}
\title{Select markers to be excluded from GWAS scan.}
\usage{
exclMarkers(snpCov, markers, allFreq, ref = NULL)
}
\arguments{
\item{snpCov}{A character vector of snpCovariates.}

\item{markers}{A matrix with marker information.}

\item{allFreq}{A numerical vector of allele frequencies of the markers in
\code{markers}. This could be computed from markers as well but it is
needed in the general algorithm so to not redo things unnecessarily it is
not redone here.}
}
\value{
A numerical vector of markers to be exluded from the GWAS scan.
}
\description{
Helper function for selecting markers to be excluded from GWAS scan.
Markers are excluded if they are identical to any of the snpCovariates
(including the snpCovariates themselves).
}
\keyword{internal}
