% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGLS.R
\name{fastGLS}
\alias{fastGLS}
\title{Fast Generalized Least Squares algoritm}
\usage{
fastGLS(y, X, Sigma, covs = NULL, nCores = NULL)
}
\arguments{
\item{y}{A numeric vector of length n of phenotypic scores. No missing
values allowed.}

\item{X}{An n x m matrix of marker-scores, n being the number of
individuals, m the number of markers. no missing values allowed.}

\item{Sigma}{An n x n covariance matrix. No missing values allowed.}

\item{covs}{An n x c matrix of covariates (NOT including an intercept).
No missing values allowed.}

\item{nCores}{A numerical value indicating the number of cores to be used by
the parallel part of the algorithm. If \code{NULL} the number of cores used
will be equal to the number of cores available on the machine - 1.}
}
\value{
A data.table with the following columns:
\itemize{
\item{\code{pValue} p-values for the GLS F-test}
\item{\code{beta} effect sizes}
\item{\code{betaSe} standard errors of the effect sizes}
\item{\code{RLR2} L_LR^2 statistics as defined in Sun et al.}
\item{\code{rn} SNP name}
}
}
\description{
Compute statistics for the Generalized Least Squares (GLS) F-test based on
the algorithm proposed by Segura (2012). Also the \eqn{R_LR^2} statistics
as in Sun (2010) is computed.
}
\references{
Segura et al. (2012) An efficient multi-locus mixed-model
approach for genome-wide association studies in structured populations.
Nature Genetics, June 2012, Vol. 44, p. 825–830.

Sun et al. (2010) Variation explained in mixed-model association
mapping. Heredity, February 2010, Vol. 105, p. 333–340.
}
\keyword{internal}
