% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gDataRiceCodedBeagle}
\alias{gDataRiceCodedBeagle}
\title{Data for vignette}
\format{An object of class \code{gData} of length 5.}
\usage{
gDataRiceCodedBeagle
}
\description{
Data used in vignette
}
\keyword{datasets}
