% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gDataTest}
\alias{gDataTest}
\title{Random test data for unit testing.}
\format{An object of class \code{gData} of length 5.}
\usage{
gDataTest
}
\description{
Random test data for unit testing.
}
\keyword{internal}
