% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASHelp.R
\name{genCtrlPVals}
\alias{genCtrlPVals}
\title{Correction of p-values based on genomic inflation}
\usage{
genCtrlPVals(pVals, nObs, nCov = 0)
}
\arguments{
\item{pVals}{A numeric vector of p-values between 0 and 1; may contain NA's.}

\item{nObs}{An integer > 0 indicating the number of individuals.}

\item{nCov}{An integer > 0 indicating the number of covariables.}
}
\value{
A list with two components:
\itemize{
\item{\code{pValues} a vector of p-values corrected by the genomic inflation
factor, with the same NA's as the input}.
\item{\code{inflation} the inflation factor}.
}
}
\description{
Correction of p-values based on the genomic inflation factor, as in Devlin
and Roeder (1999). It is assumed that the p-values come from an F-test with
df1 = 1 and df2 = nObs - nCov - 2.
}
\references{
Devlin, B. and Roeder K. (1999) Genomic control for association
studies. Biometrics, December 1999, Vol. 55(4), p. 997-1004.
}
\keyword{internal}
