% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASHelp.R
\name{getSNPsInRegionSufLD}
\alias{getSNPsInRegionSufLD}
\title{get the SNPs close to a given SNP with sufficient LD}
\usage{
getSNPsInRegionSufLD(gData, snp, regionSize = 5000, minR2 = 0.5)
}
\arguments{
\item{gData}{An object of class gData with at least the map and markers
included.}

\item{snp}{An integer indicating the index of the reference SNP within
the map.}

\item{regionSize}{A numerical value indicating the size of the region on
the chromosome in which to look for SNPs.}

\item{minR2}{A numerical value between 0 and 1 indicating the minimum
LD (in terms of r^2) that the SNPs should have with the reference SNP.}
}
\value{
An integer vector with indices of the SNPs that are within the
given \code{regionSize} and have a minimum LD with the reference SNP.
}
\description{
\code{getSNPsInRegionSufLD} extracts the SNPs from a map file that are
within a given distance of a reference SNP (on either side). Only those SNPs
that are in sufficient linkage disequilibrium (LD) with the reference SNP
are returned.
}
\keyword{internal}
