% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeMarkers.R
\name{imputeBeagle}
\alias{imputeBeagle}
\title{Helper function for imputation using beagle.}
\usage{
imputeBeagle(markersRecoded, map)
}
\description{
Helper function for imputation using beagle.
}
\keyword{internal}
