% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Create a manhattan plot}
\usage{
manhattanPlot(xValues, yValues, map, xLab = "Chromosomes",
  yLab = expression(-log[10](p)), xSig = integer(),
  xEffects = integer(), colPalette = rep(c("black", "grey50"), 50),
  yThr = NULL, signLwd = 0.6, ..., output = TRUE)
}
\arguments{
\item{xValues}{A vector of cumulative marker positions.}

\item{yValues}{A vector of LOD-scores.}

\item{map}{A data.frame with at least the columns chr, the number of the
chromosome and cumPos, the cumulative position of the snp the cumulative
position of the snp starting from the first chromosome.}

\item{xLab}{A character string, the x-axis label.}

\item{yLab}{A character string, the y-axis label.}

\item{xSig}{A vector of integers, indicating which components in the vectors
xValues and yValues are significant.}

\item{xEffects}{A vector of integers, indicating which components in the
vector xValues correspond to a real (known) effect.}

\item{colPalette}{A color palette used for plotting.}

\item{yThr}{A numerical value for the LOD-threshold.}

\item{signLwd}{A numerical value giving the thickness of the
points that are false/true positives/negatives.}

\item{...}{Other graphical parameters passed on to the actual plot function.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned.}
}
\value{
A LOD-profile with LOD-scores per snip. Markers declared significant
get a red dot, markers with a real effect get a blue dot. If both significant
and real effects are given false positives get an orange dot, true negatives
a yellow dot and true positives a green dot.
}
\description{
Given vectors of marker positions and corresponding LOD-scores plot a
LOD-profile. Significant markers can be highlighted with red dots. If there
are previously known marker effect also false positives and true negatives
will be marked.
}
\keyword{internal}
