% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixRoot}
\alias{matrixRoot}
\title{Compute square root of a symmetric, positive definite matrix}
\usage{
matrixRoot(x)
}
\arguments{
\item{X}{A symmetric, positive definite matrix.}
}
\value{
A matrix Y such that \eqn{Y^2 = X}.
}
\description{
Given a symmetric, positive definite matrix X a matrix Y is computed such
that \eqn{Y^2 = X}. Computation is done using eigendecomposition of X.
}
\keyword{internal}
