% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlPlot.R
\name{qtlPlot}
\alias{qtlPlot}
\title{Draw a qtl plot}
\usage{
qtlPlot(dat, map, normalize = FALSE, sortData = FALSE,
  binPositions = NULL, printVertGrid = TRUE, yLab = "Traits",
  exportPptx = FALSE, pptxName = "", output = TRUE, ...)
}
\arguments{
\item{dat}{A data.frame with QTL data to be plotted}

\item{map}{A data.frame with at least the columns \code{chr}, the number of
the chromosome and \code{pos}, the position of the SNP on the chromosome.
These are used for calculating the physical limits of the chromosome.}

\item{normalize}{Should the snpEffect be normalized?}

\item{sortData}{Should the data be sorted before plotting? Either
\code{FALSE} if no sorting should be done or a character indicating the data
column on which the data should be sorted.}

\item{binPositions}{An optional data.frame containing at leasts two columns,
chr and pos. Vertical lines are plotted at those positions.}

\item{printVertGrid}{Should default vertical grid lines be plotted.}

\item{yLab}{A character string, the y-axis label.}

\item{exportPptx}{Should the plot be exported to a .pptx file?}

\item{pptxName}{A character string, the name of the .pptx file to which the
plot is exported. Ignored if exportPptx = \code{FALSE}.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned.}

\item{...}{Further parameter.}

\item{chromosome}{A character string indicating the data column containing
the chromosome number.}

\item{trait}{A character string indicating the data column containing the
trait name.}

\item{snpEffect}{A character string indicating the data column containing
the SNP effect.}

\item{snpPosition}{A character string indicating the data column containing
the position of the SNP on the chromosome.}
}
\value{
A ggplot object is returned and plotted on screen. Optionally the
plot is exported to pptx as well.
}
\description{
Visualisation of the final set of QTLs selected after the multi-trait GWAS.
}
\details{
Each horizontal line contains QTLs of one trait, phenotypic trait or
trial. Option: Vertical white lines can indicate chromosome
subdivision, genes of interest, known QTLs, etc. Circle diameters are
proportional to the absolute value of allelic effect. Colors indicate the
direction of the effect: green when the allele increases the trait value,
and blue when it decreases the value.
}
\references{
Millet et al. (2016) Genome-wide analysis of yield in Europe:
Allelic effects vary with drought and heat scenarios. Plant Physiology,
October 2016, Vol. 172, p. 749–764
}
\keyword{internal}
