% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reduceKinship}
\alias{reduceKinship}
\title{Reduce the kinship matrix}
\usage{
reduceKinship(K, nPca)
}
\arguments{
\item{nPca}{An integer, the number of eigenvectors used for reducing the
kinship matrix.}
}
\value{
The reduced kinship matrix
}
\description{
The kinship matrix is reduced using nPca eigenvectors of K.
}
\keyword{internal}
