% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenGWAS.R
\docType{package}
\name{statgenGWAS-package}
\alias{statgenGWAS}
\alias{statgenGWAS-package}
\title{statgenGWAS: Genome Wide Association Studies}
\description{
Fast single trait Genome Wide Association Studies (GWAS) following the method described in Kang et al. (2010), \doi{10.1038/ng.548}. One of a series of statistical genetic packages for streamlining the analysis of typical plant breeding experiments developed by Biometris.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometris.github.io/statgenGWAS/index.html}
  \item \url{https://github.com/Biometris/statgenGWAS/}
  \item Report bugs at \url{https://github.com/Biometris/statgenGWAS/issues}
}

}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Authors:
\itemize{
  \item Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})
  \item Willem Kruijer (\href{https://orcid.org/0000-0001-7179-1733}{ORCID})
}

Other contributors:
\itemize{
  \item Fred van Eeuwijk (\href{https://orcid.org/0000-0003-3672-2921}{ORCID}) [contributor]
  \item Martin Boer (\href{https://orcid.org/0000-0002-1879-4588}{ORCID}) [contributor]
  \item Marcos Malosetti (\href{https://orcid.org/0000-0002-8150-1397}{ORCID}) [contributor]
  \item Daniela Bustos-Korts (\href{https://orcid.org/0000-0003-3827-6726}{ORCID}) [contributor]
  \item Emilie Millet (\href{https://orcid.org/0000-0002-2913-4892}{ORCID}) [contributor]
  \item Joao Paulo (\href{https://orcid.org/0000-0002-4180-0763}{ORCID}) [contributor]
  \item Maikel Verouden (\href{https://orcid.org/0000-0002-4893-3323}{ORCID}) [contributor]
  \item Ron Wehrens (\href{https://orcid.org/0000-0002-8798-5599}{ORCID}) [contributor]
  \item Choazhi Zheng (\href{https://orcid.org/0000-0001-6030-3933}{ORCID}) [contributor]
}

}
\keyword{internal}
