% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA.R
\name{EMMA}
\alias{EMMA}
\title{Compute REML estimates of variance components using EMMA algorithm.}
\usage{
EMMA(gData, trait, environment, K = NULL, covar = NULL,
  snpName = NULL, Z = NULL, nGrids = 100, lLim = -10, uLim = 10,
  eps = .Machine$double.eps^0.25)
}
\arguments{
\item{gData}{An object of class gData containing at least a data.frame
\code{pheno}. If \code{K} is not supplied a matrix \code{kinship} should be
in \code{gData}. If covariates are included then a data.frame covar is
needed as wel and if an extra snp is to be included as covariate (defined
in \code{snpName}) then a data.frame \code{markers} is also needed. Missing
values in \code{pheno} are allowed but will be excluded from the
calculations.}

\item{trait}{A trait for which to estimate variance components. This can be
either numeric index or character name of a column in \code{pheno}.}

\item{environment}{An environment for which to estimate variance components.
This can be either numeric index or character name of a list item in
\code{pheno}.}

\item{K}{An optional kinship matrix. If \code{NULL} then matrix
\code{kinship} in \code{gData} is used. If both \code{K} is provided and
\code{gData} contains a matrix \code{kinship} then \code{K} is used.}

\item{covar}{An optional vector of covariates taken into account when
estimating variance components. These can be either numeric indices or
character names of columns in \code{covar} in \code{gData}. If \code{NULL}
no covariates are used.}

\item{snpName}{An optional character string of a marker in \code{markers} in
\code{gData} to be included as covariate. If used the \code{gData} object
should contain a data.frame \code{markers}.}

\item{Z}{An optional incidence matrix mapping each observed phenotype to
one of inbred strains.}

\item{nGrids}{An integer indicating the number of intervals used for local
optimisation within the algorithm.}

\item{lLim}{A numerical value indicating the lower limit of the interval over
which optimisating is done.}

\item{uLim}{A numerical value indicating the upper limit of the interval over
which optimisating is done.}

\item{eps}{A numerical value used as computational tolerance in the
algorithm.}
}
\value{
A list with two components:
\itemize{
\item{\code{varcomp} a vector of genetic variance Vg and residual variance
Ve}
\item{\code{vcovMatrix} The variance covariance matrix corresponding to
the computed variances.}
}
}
\description{
Using the EMMA algorithm as in Kang et al. (2008) to compute REML estimates
of genetic and residual variance components.
}
\references{
Kang et al. (2008) (Efficient Control of Population Structure in
Model Organism Association Mapping. Genetics, March 2008, Vol. 178, no. 3,
p. 1709-1723
}
\keyword{internal}
