% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGData.R
\name{plot.gData}
\alias{plot.gData}
\title{Plot function for the class \code{GWAS}}
\usage{
\method{plot}{gData}(x, ..., plotType = c("genMap"), title = NULL, output = TRUE)
}
\arguments{
\item{x}{An object of class \code{gData}.}

\item{...}{Further arguments to be passed on to the actual plotting
functions.}

\item{plotType}{A character string indicating the type of plot to be made.
One of "genMap".}

\item{title}{A character string, the title of the plot.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE}, only a list of ggplot objects is invisibly returned.}
}
\description{
Creates a plot of an object of S3 class \code{GWAS}. The following types of
plot can be made:
\itemize{
\item{a manhattan plot, i.e. a plot of LOD-scores per SNP}
}
}
