% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGWAS.R
\name{summary.GWAS}
\alias{summary.GWAS}
\title{Summary function for the class \code{GWAS}}
\usage{
\method{summary}{GWAS}(object, ..., trials = NULL)
}
\arguments{
\item{object}{An object of class \code{GWAS}.}

\item{...}{Not used.}

\item{trials}{A vector of strings or numeric indices indicating for
which trials the summary should be made. If \code{NULL}, a summary is
made for all trials.}
}
\description{
Gives a summary for an object of S3 class \code{GWAS}.
}
